feat(rust): Schritt 4d/1 — EventSub-Ingress (Dispatch, Handler, Telemetrie)
POST /eventsub/dispatch (Bridge-Vertrag) in tb-internal-api, Dispatcher
mit persistentem Message-Dedup (600s) + Inbox-Enqueue für Core-Typen,
Inbox-Handler (online/offline/update mit Business-Effect-Guards 7d,
Offline-Throttle 120s), Telemetrie-Inserts (prod-verifizierte Typen),
Raid/Moderate via EventSubHooks. tb-bot-Wiring mit Noop-Hooks.

Fix: streamers_crud-Exists-Check dekodierte SELECT 1 als i64 → 500er
bei doppeltem Add (latent seit 3b, Tests liefen damals ohne Test-DB).

+++ rust/crates/tb-analytics/src/streamers_crud.rs
@@ -60,8 +60,8 @@ pub async fn add_streamer(
 ) -> Result<AddStreamerResult, sqlx::Error> {
     let login = login.to_lowercase();
 
-    // Prüfen ob bereits aktiv
-    let exists: Option<(i64,)> = sqlx::query_as(
+    // Prüfen ob bereits aktiv (`SELECT 1` ist INT4 — als i32 dekodieren)
+    let exists: Option<(i32,)> = sqlx::query_as(
         "SELECT 1 FROM twitch_streamers WHERE LOWER(twitch_login) = LOWER($1) AND archived_at IS NULL",
     )
     .bind(&login)
